Compartmentalise project structure under src
