+++ src/utils/obsidian.ts
@@ -1,0 +1,5 @@
+import { FileSystemAdapter} from "obsidian";
+
+export function getVaultPath(adapter: FileSystemAdapter): string {
+  return adapter.getBasePath();
+}
